Change to CMake project
