--- Mebius.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33627.172
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Mebius", "Mebius\Mebius.vcxproj", "{4BA032B5-1A11-42D6-8360-E1CEB96CB39C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4BA032B5-1A11-42D6-8360-E1CEB96CB39C}.Debug|x86.ActiveCfg = Debug|Win32
-		{4BA032B5-1A11-42D6-8360-E1CEB96CB39C}.Debug|x86.Build.0 = Debug|Win32
-		{4BA032B5-1A11-42D6-8360-E1CEB96CB39C}.Release|x86.ActiveCfg = Release|Win32
-		{4BA032B5-1A11-42D6-8360-E1CEB96CB39C}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D620FFF2-4C7C-4C26-9F9D-330656C0F72C}
-	EndGlobalSection
-EndGlobal
