--- Mebius/framework.h
@@ -1,5 +1,0 @@
-﻿#pragma once
-
-#define WIN32_LEAN_AND_MEAN             // Windows ヘッダーからほとんど使用されていない部分を除外する
-// Windows ヘッダー ファイル
-#include <windows.h>
