--- Mebius/pch.cpp
@@ -1,5 +1,0 @@
-﻿// pch.cpp: プリコンパイル済みヘッダーに対応するソース ファイル
-
-#include "pch.h"
-
-// プリコンパイル済みヘッダーを使用している場合、コンパイルを成功させるにはこのソース ファイルが必要です。
