+++ include/framework.h
@@ -1,0 +1,5 @@
+﻿#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Windows ヘッダーからほとんど使用されていない部分を除外する
+// Windows ヘッダー ファイル
+#include <windows.h>
