+++ src/pch.cpp
@@ -1,0 +1,5 @@
+﻿// pch.cpp: プリコンパイル済みヘッダーに対応するソース ファイル
+
+#include "pch.h"
+
+// プリコンパイル済みヘッダーを使用している場合、コンパイルを成功させるにはこのソース ファイルが必要です。
